+++ packages/catalyst_builder/example/lib/example.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_builder/catalyst_builder.dart';
+import 'package:catalyst_builder_contracts/catalyst_builder_contracts.dart';
 
 import './src/manually_wired_service.dart';
 
@@ -9,7 +9,6 @@ export 'example.catalyst_builder.g.dart';
 @Preload()
 @GenerateServiceProvider(
   providerClassName: 'ExampleProvider',
-  includePackageDependencies: true,
 )
 @ServiceMap(services: {
   ManuallyWiredServiceImplementation: Service(