+++ packages/catalyst_builder/lib/src/builder/dto/dto.dart
@@ -1,4 +1,3 @@
-export 'config.dart';
 export 'constructor_arg.dart';
 export 'extracted_service.dart';
 export 'inject_annotation.dart';