+++ packages/catalyst_builder/lib/src/builder/dto/entrypoint.dart
@@ -1,27 +1,27 @@
 /// Represents a simple symbol reference.
 class Entrypoint {
-  /// The name of the service provider
+  /// The name of the service provider class.
   final String providerClassName;
-
-  /// Should the service provider include services from dependencies
-  final bool includePackageDependencies;
 
   /// The assetId
   final Uri assetId;
 
+  /// The name of the plugin class.
+  final String pluginClassName;
+
   /// Instantiate a new entrypoint
   const Entrypoint({
     required this.providerClassName,
-    required this.includePackageDependencies,
     required this.assetId,
+    required this.pluginClassName,
   });
 
   /// Creates a new instance from the result of [toJson].
   factory Entrypoint.fromJson(Map<String, dynamic> json) {
     return Entrypoint(
       providerClassName: json['entrypoint'].toString(),
-      includePackageDependencies: json['includePackageDependencies'] as bool,
       assetId: Uri.parse(json['assetId'].toString()),
+      pluginClassName: json['pluginClassName'].toString(),
     );
   }
 
@@ -29,8 +29,8 @@ class Entrypoint {
   Map<String, dynamic> toJson() {
     return {
       'entrypoint': providerClassName,
-      'includePackageDependencies': includePackageDependencies,
       'assetId': assetId.toString(),
+      'pluginClassName': pluginClassName,
     };
   }
 }