+++ packages/catalyst_builder/lib/src/builder/generator/service_provider/fields/applied_plugins.dart
@@ -1,0 +1,11 @@
+import 'package:code_builder/code_builder.dart' as cb;
+
+import '../../symbols.dart';
+
+/// Template for _appliedPlugins
+final appliedPluginsTemplate = cb.Field((f) {
+  f
+    ..name = appliedPlugins$.symbol
+    ..modifier = cb.FieldModifier.final$
+    ..assignment = cb.literalList([], serviceProviderPluginT).code;
+});