+++ packages/catalyst_builder/lib/src/builder/generator/service_provider/fields/fields.dart
@@ -1,5 +1,7 @@
+export 'applied_plugins.dart';
 export 'booted.dart';
 export 'expose_map.dart';
 export 'known_services.dart';
+export 'preloaded_types.dart';
 export 'service_instances.dart';
 export 'services_by_tag.dart';