+++ packages/catalyst_builder/lib/src/builder/generator/service_provider/methods/resolve_or_get_parameter.dart
@@ -32,6 +32,7 @@ final resolveOrGetParameterTemplate = cb.Method((m) {
 
   m
     ..name = resolveOrGetParameter$.symbol
+    ..annotations.add(cb.refer('override'))
     ..requiredParameters.addAll([
       cb.Parameter((p) => p
         ..name = paramRequiredBy$.symbol!