+++ packages/catalyst_builder_contracts/lib/src/annotation/annotation.dart
@@ -1,4 +1,5 @@
 export 'generate_service_provider.dart';
+export 'generate_service_provider_plugin.dart';
 export 'inject.dart';
 export 'preload.dart';
 export 'service.dart';