+++ packages/catalyst_builder_contracts/lib/src/annotation/generate_service_provider.dart
@@ -4,14 +4,17 @@ class GenerateServiceProvider {
   /// service provider.
   final String providerClassName;
 
-  /// Set this property to true if you like to include also the services
-  /// provided of your dependencies. Otherwise only the services in your
-  /// root package will be registered.
-  final bool includePackageDependencies;
+  /// Set this property for changing the class name of the generated
+  /// service provider plugin.
+  ///
+  /// If you're developing a third party plugin you need to set this property
+  /// otherwise the plugin class is not public since the default value is
+  /// prefixed by _;
+  final String pluginClassName;
 
   /// Mark this file as the entry point for the service container
   const GenerateServiceProvider({
     this.providerClassName = 'DefaultServiceProvider',
-    this.includePackageDependencies = false,
-  });
+    String? pluginClassName,
+  }) : pluginClassName = pluginClassName ?? "_${providerClassName}Plugin";
 }