+++ packages/catalyst_builder_contracts/lib/src/annotation/generate_service_provider_plugin.dart
@@ -1,0 +1,11 @@
+/// Mark the file which contains this annotation as the entry point.
+class GenerateServiceProviderPlugin {
+  /// Set this property for setting the class name of the generated
+  /// service provider plugin.
+  final String pluginClassName;
+
+  /// Mark this file as the entry point for the plugin
+  const GenerateServiceProviderPlugin({
+    required this.pluginClassName,
+  });
+}