+++ packages/catalyst_builder/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 5.0.0-dev.1
+
+We need a total makeover of the underlying architecture.
+Services from packages you depend on are no longer resolved automatically!
+
+More background information is available in the [Wiki](https://github.com/mintware-de/catalyst_builder/wiki/v5) 
+
+### Changes
+- `includePackageDependencies` was removed
+
 ## 4.3.2
 
 ### Changes
@@ -9,26 +19,29 @@
 ### Changes
 
 - Removed verbose logging
+- Removed the caching system
 
 ## 4.3.0
 
 ### Changes
 
 #### Migrated the ServiceProviderBuilder to post_process_builders
+
 Since the ServiceProviderBuilder does not rely on the real outputs of the previous builders but the cache files,
 this is a much better solution since a PostProcessBuilders does not process all .dart files.
 
 ### Fixes
 
 #### Skip unprocessable files
+
 This will fix the following error when using catalyst_builder with Flutter:
+
 ```plain
 [SEVERE] catalyst_builder:preflight on package:sky_engine/html/html_dart2js.dart:
 
 This builder requires Dart inputs without syntax errors.
 ...
 ```
-
 
 ## 4.2.1
 
@@ -41,8 +54,9 @@ This builder requires Dart inputs without syntax errors.
 ### Changes
 
 #### Project separation
+
 To improve the maintainability of your projects, we decided to separate the less frequently changing parts - such as
-annotations, abstractions like the base class for the ServiceProvider, and exceptions (`catalyst_builder_contracts`) - 
+annotations, abstractions like the base class for the ServiceProvider, and exceptions (`catalyst_builder_contracts`) -
 from the more frequently changing parts, like the implementation of the builder itself (`catalyst_builder`).
 
 This means that a breaking change in the `catalyst_builder` package does not force you to update projects where you only
@@ -54,7 +68,8 @@ use the annotations.
 
 #### Cache location
 
-You're now able to set a custom cache location by setting `catalyst_builder: { cacheDir: 'a/cache/path' }` property in the pubspec.yaml
+You're now able to set a custom cache location by setting `catalyst_builder: { cacheDir: 'a/cache/path' }` property in
+the pubspec.yaml
 
 Take a look in [pubspec.yaml](example/pubspec.yaml) for an example.